+++ ErsatzTV.FFmpeg/Option/CopyTimestampInputOption.cs
@@ -1,0 +1,21 @@
+﻿using ErsatzTV.FFmpeg.Environment;
+
+namespace ErsatzTV.FFmpeg.Option;
+
+public class CopyTimestampInputOption : IInputOption
+{
+    public IList<EnvironmentVariable> EnvironmentVariables => Array.Empty<EnvironmentVariable>();
+    public IList<string> GlobalOptions => Array.Empty<string>();
+
+    public IList<string> InputOptions(InputFile inputFile) => new List<string> { "-copyts" };
+
+    public IList<string> FilterOptions => Array.Empty<string>();
+    public IList<string> OutputOptions => Array.Empty<string>();
+    public FrameState NextState(FrameState currentState) => currentState;
+
+    public bool AppliesTo(AudioInputFile audioInputFile) => false;
+
+    public bool AppliesTo(VideoInputFile videoInputFile) => true;
+
+    public bool AppliesTo(ConcatInputFile concatInputFile) => false;
+}